Commited V2.4.0D (June 17 2011 release) from http://renaud91.free.fr/MetaDataExtractor/
Changes in 2.4.0d (17 June 2011):
Following the remark of Lars �ke Vinberg, I change the code in order to use the foreach optimisation for the Directory and Metadata classes.
I keept the old GetIterator method for compatibility.

Changes in 2.4.0c (14 January 2011):
Corrected a bug in ExifReader.cs where a space was preventing introspection of OlympusDirectory to work (thanks to Zvika Gart).
Changed the code for method StoreThumbnailBytes in order to use Buffer.BlockCopy instead of the manual byte/byte copy (thanks to Zvika Gart).
Added some docmentation in Rationnal class in order to warn about IntValue and other methods when denominator is 0 (thanks to Steven Jeuris).

Changes in 2.4.0b (18 April 2009):
Corrected a bug in IPTC reading when date was wrong (thanks to Mathieu).
Corrected a bug in method that was supposed to filter image (occured when there was only one file).

Changes in 2.4.0a (30 Juin 2008):
Changed back Dispose call to Close call since Dispose does not always close Stream.
Added some Close where it was needed (Thanks to X B�hlmann)
Added some Canon tags (Thanks to X B�hlmann)

Changes in 2.4.0 (22 Mai 2008):
One big change in this version is the way resources are handled. I was fed up with the C# .resources that does not work well.
So now, a resource file is a simple txt file that can be found in the resource folder.
You can translate them if you want and add them to the project. Simply create a directory (for example resource\fr) and put the file in it. DO NOT FORGET to make a right click and add to project as embeded resource.
The behaviour is as follow : if resource is in the langage folder, then it will be used, if not it will use the default resource file (the one at the root of the resource directory).
With this system there should not be any more error regarding resource file (I hope ...).
Impacted change from beta version of Drew, so this version can read metadata from RAW files (*.raw, *.cr2, *.crw). Note that this does not work with all raw format (it works for me with the 5D raw, but not with the 10D raw)
Changed the Default namespace to MetaDataExtractor (instead of com).
Corrected the bug found by TIMOTHY regarding the GetFocalPlaneYResolutionDescription in ExifDirectory.
Corrected a bug where some date may not be parsed correcly.
Project is compiled for .NET 2.0 but is to be used with Visual Studio Free Edition 2008. It is fully compatible with 3.0 and 3.5 .NET version
Added a new XML output format (use asXml2 instead of asXml). I use this version and found it more practical.
Changed the API in order to be more object
Added a few support for Trace and Debug (using System.Diagnostics)